% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{make_gsea_input}
\alias{make_gsea_input}
\title{prepare GSEA input files from GSE raw data}
\usage{
make_gsea_input(matrix_file, soft_file, output_dir = ".", gene = NULL)
}
\arguments{
\item{matrix_file}{string. path to GSE series matrix file}

\item{soft_file}{string. path to GSE SOFT file}

\item{output_dir}{string, path to output directory}

\item{gene}{character. HUGO gene symbol(s) of your interested gene(s)}
}
\value{
\code{NULL}
}
\description{
You can use first example if you already have a \href{http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Phenotype_Labels}{phenotype labels file}, or use second example if you are interested in gene sets co-expressing with certain gene(s).
}
\details{
both \code{NA} and \code{Inf} are not allowed in input files (\code{.txt} & \code{.cls}).
}
\examples{
\dontrun{
  make_gsea_input(
    system.file('extdata/GSE51280_series_matrix.txt.gz', package = 'rGEO'),
    system.file('extdata/GSE51280_family.soft.gz', package = 'rGEO')
  )

  make_gsea_input(
    system.file('extdata/GSE19161_series_matrix.txt.gz', package = 'rGEO'),
    system.file('extdata/GSE19161_family.soft.gz', package = 'rGEO'),
    '.', 'EIF4G2'
  )
}

}
