% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{read_chip}
\alias{read_chip}
\title{read .chip file info a \code{chip}}
\usage{
read_chip(file)
}
\value{
tibble. identical to \link{read_gse_soft}
}
\description{
read .chip file info a \code{chip}
}
\seealso{
\link{write_chip}
}
