% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{read_txt}
\alias{read_txt}
\title{read .txt file info a \code{matrix}}
\usage{
read_txt(file)
}
\value{
tibble. identical to \code{\link[rGEO:read_gse_matrix]{rGEO::read_gse_matrix()}}
}
\description{
read .txt file info a \code{matrix}
}
\seealso{
\link{write_txt}
}
