% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{write_chip}
\alias{write_chip}
\title{write GSEA microarray chip annotation input file}
\usage{
write_chip(chip, path)
}
\arguments{
\item{chip}{tibble. usually the return value of \code{\link[rGEO:read_gse_soft]{rGEO::read_gse_soft()}}, \code{NULL} is okay.}

\item{path}{string. path to output file}
}
\value{
logical scalar
}
\description{
write GSEA microarray chip annotation input file
}
\section{to do}{
 fetch description from hgnc later
}

\examples{
`NULL`

}
\seealso{
\link{read_chip}
}
