% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{write_cls}
\alias{write_cls}
\title{write GSEA phenotype data input file}
\usage{
write_cls(phenotype, path)
}
\arguments{
\item{phenotype}{tibble. usually the return value of \link{make_phenotype}}

\item{path}{string. path to output file}
}
\value{
logical scalar
}
\description{
only support continuous phenotype for now
}
\details{
Although GSEA doesn't require every variable in a .cls file to be the same length, I do. Since they should be used with the same expression data (such as GSE008.txt + GSE008.chip + GSE008.cls, on the other hand, let GSE006.txt and GSE007.txt use different variable in the same .cls file is definitely insane)
}
\examples{
`NULL`

}
