% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{write_txt}
\alias{write_txt}
\title{write GSEA expression dataset input file}
\usage{
write_txt(matrix, path)
}
\arguments{
\item{matrix}{tibble. usually the return value of \code{\link[rGEO:read_gse_matrix]{rGEO::read_gse_matrix()}}, \code{NULL} is okay.}

\item{path}{string. path to output file}
}
\value{
logical scalar
}
\description{
write GSEA expression dataset input file
}
\examples{
NULL

}
\seealso{
\link{read_txt}
}
